sdk: refactor: move address metadata and pfs codecs
Some codecs, types and utilities were spread across services, transport,
transfers and messages types and utilities modules. This moves all the
relevant bits into services's types & utils, in order to better reuse
them across depending modules.
